+++ CHANGELOG.md
@@ -51,6 +51,7 @@ Unreleased
 - Fix bug that `EventListener::stop` did not actually stop any async tasks.
 - Add function `EventListenerCfg::async_tick` to switch the ticker from Sync-Port to Async-Ports. Note that `EventListenerCfg::tick_interval` is still necessary to enable any ticker.
 - Dont start the Sync-Port worker if there are no sync ports (note that `async_tick(false)`(the default) also counts as a sync port).
+- Fix accidental always inclusion of `crossterm` dependency, even if `crossterm` feature was disabled. (since [2.2.0](#220))
 
 ## 2.4.0