+++ CHANGELOG.md
@@ -52,6 +52,7 @@ Unreleased
 - Add function `EventListenerCfg::async_tick` to switch the ticker from Sync-Port to Async-Ports. Note that `EventListenerCfg::tick_interval` is still necessary to enable any ticker.
 - Dont start the Sync-Port worker if there are no sync ports (note that `async_tick(false)`(the default) also counts as a sync port).
 - Fix accidental always inclusion of `crossterm` dependency, even if `crossterm` feature was disabled. (since [2.2.0](#220))
+- Change `PollAsync::poll` to take `&mut self` instead of `&self`.
 
 ## 2.4.0